Add support for additional controller mappings
SDL2 (and love2d) has a limited number of gamepad mappings builtin. I
found that the controllers I tried to use when I first played didn't
work out of the box.

The SDL_GameControllerDB database is a crowdsourced database of a large
number of game controller default mappings used by a number of games so
that more gamepads work out of the box. I've imported a copy of the DB
into vendor, and configured the game to use it; it's distributed under
the zlib license.